% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_class_01_GEMA.dancefloor.R
\docType{class}
\name{GEMA.dancefloor,-class}
\alias{GEMA.dancefloor,-class}
\alias{GEMA.dancefloor-class}
\title{S4 Class GEMA.dancefloor}
\description{
This class is used for objects that define one dancefloor.
}
\details{
Use function \code{\link[GEMATariff:dancefloor]{dancefloor}} to create objects of this class.
}
\section{Slots}{

\describe{
\item{\code{space}}{Numeric, room size for the dancefloor in square meters.}

\item{\code{days}}{Integer number between 1 and 7,
      number of days the club is opened per week.}

\item{\code{admission}}{Numeric,
      highest ticket price for the dancefloor. If prices vary over the week,
use the average of highest prices, respectively.}
}}
\examples{
myDancefloor <- dancefloor(space=200, days=3, admission=6)
}
\keyword{classes}

