% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_class_03_GEMA.gig.R
\docType{class}
\name{GEMA.gig,-class}
\alias{GEMA.gig,-class}
\alias{GEMA.gig-class}
\title{S4 Class GEMA.gig}
\description{
This class is used for objects that define a live concert.
}
\details{
Use function \code{\link[GEMATariff:gig]{gig}} to create objects of this class.
}
\section{Slots}{

\describe{
\item{\code{gross}}{Numeric, the total gross receipts for the concert.}

\item{\code{guests}}{Numeric, number of people attending the concert.}

\item{\code{admission}}{Numeric, highest ticket price for the concert.}

\item{\code{musicians}}{Logical,
      whether the music ist mostly perfomed by live musicians (in contrast to
mostly by using physical recording media).}

\item{\code{addIncome}}{Logical,
      whether you have additional income through advertising etc.}

\item{\code{public}}{Logical, whether the concert was either public with no admission fee,
      or an invite-only
event for a strictly limited amount of guests (e.g., a promotional event). If \code{TRUE},
      \code{fee}
and \code{production} become relevant, and \code{admission} must be 0.}

\item{\code{fee}}{Numeric, the amount of money payed to the artists,
      relevant if \code{public=TRUE}.}

\item{\code{production}}{Numeric, the directly associated production costs,
      relevant if \code{public=TRUE}.}

\item{\code{charity}}{Logical, whether the event is of non-profit/charitable character.}

\item{\code{rcs}}{Logical, whether the event is of religious, cultural,
      or social character.}

\item{\code{promoYoung}}{Logical, whether this event is only promoting young musicians.}

\item{\code{lumpSumContract}}{Numeric,
      one of the following to declare it you have a yearly contract:
\describe{
  \item{0}{No yearly contract or not more than 15 events a year.}
  \item{16}{Yearly contract and at least 16 events a year.}
  \item{31}{Yearly contract and at least 31 events a year.}
}}

\item{\code{plusMedium}}{Logical,
      whether in addition to the live music there is also music from
physical recording media (e.g., CDs) being played.}
}}

\examples{
myGig <- gig()
}
\keyword{classes}
