% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_class_04_GEMA.invoice.R
\docType{class}
\name{GEMA.invoice,-class}
\alias{GEMA.invoice,-class}
\alias{GEMA.invoice-class}
\title{S4 Class GEMA.invoice}
\description{
This class is inherited by other invoice classes and not standalone.
}
\section{Slots}{

\describe{
\item{\code{factors}}{A named list of factors that affect the end result:
\describe{
  \item{GVL}{Fees for GVL.}
  \item{VAT}{Value added tax.}
  \item{launch}{Reduction factor for market introduction of the new tariffs.}
}}

\item{\code{categories}}{A named list of price categories:
\describe{
  \item{space}{Space category (in 100 square meter steps).}
  \item{admission}{Admission category (in 2€ steps).}
}}

\item{\code{tariff}}{A named list containing the actual monthly tariff net sums without reductions:
\describe{
   \item{GEMA}{GEMA net tariff.}
   \item{GVL}{GVL net tariff.}
 }}

\item{\code{invoice}}{A named list containing the final invoice sums:
\describe{
   \item{GEMA}{Final GEMA net tariff.}
   \item{GVL}{Final GVL net tariff.}
   \item{net}{The net sum.}
   \item{VAT}{The applicable VAT tax amount.}
   \item{gross}{The gross sum.}
 }}
}}

\keyword{classes}
