% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_class_05_GEMA.invoice_m_cd.R
\docType{class}
\name{GEMA.invoice_m_cd,-class}
\alias{GEMA.invoice_m_cd,-class}
\alias{GEMA.invoice_m_cd-class}
\title{S4 Class GEMA.invoice_m_cd}
\description{
This class is used for objects that are returned by methods like \code{\link[GEMATariff:GEMA_m_cd]{GEMA_m_cd}}.
}
\section{Slots}{

\describe{
\item{\code{factors}}{A named list of factors that affect the end result:
\describe{
  \item{GVL}{Fees for GVL.}
  \item{VAT}{Value added tax.}
  \item{launch}{Reduction factor for market introduction of the new tariffs.}
}}

\item{\code{categories}}{A named list of price categories:
\describe{
  \item{space}{Space category (in 100 square meter steps).}
  \item{admission}{Admission category (in 2€ steps).}
}}

\item{\code{dancefloor}}{An object of class \code{\link[GEMATariff]{GEMA.dancefloor-class}}.}

\item{\code{tariff}}{A named list containing the actual monthly tariff net sums without reductions:
\describe{
   \item{GEMA}{GEMA net tariff.}
   \item{GVL}{GVL net tariff.}
 }}

\item{\code{reduction}}{A named list of applicable reductions for the montly GEMA net tariff:
\describe{
   \item{tariff2012}{The original old tariff net sum (as of 2012),
      to calculate the launch reduction.}
   \item{tariffdiff}{Difference between the old tariff (plus 6.5\%) and the new one.}
   \item{launch}{The amount of reduction applicable (applying \code{launch}).}
   \item{discount}{Other global discount ("Gesamtvertragsnachlass").}
 }}

\item{\code{invoice}}{A named list containing the final invoice sums:
\describe{
   \item{GEMA}{Final GEMA net tariff.}
   \item{GVL}{Final GVL net tariff.}
   \item{net}{The net sum.}
   \item{VAT}{The applicable VAT tax amount.}
   \item{gross}{The gross sum.}
 }}
}}

\keyword{classes}
