% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_class_07_GEMA.invoice_u_k.R
\docType{class}
\name{GEMA.invoice_u_k,-class}
\alias{GEMA.invoice_u_k,-class}
\alias{GEMA.invoice_u_k-class}
\title{S4 Class GEMA.invoice_u_k}
\description{
This class is used for objects that are returned by methods like \code{\link[GEMATariff:GEMA_u_k]{GEMA_u_k}}.
}
\section{Slots}{

\describe{
\item{\code{factors}}{A named list of factors that affect the end result:
\describe{
  \item{GVL}{Fees for GVL.}
  \item{VAT}{Value added tax.}
  \item{launch}{Reduction factor for market introduction of the new tariffs.}
}}

\item{\code{categories}}{A named list of price categories:
\describe{
  \item{guests}{Guests category (in 150 guests steps).}
}}

\item{\code{gig}}{An object of class \code{\link[GEMATariff]{GEMA.gig-class}}.}

\item{\code{tariff}}{A named list containing the actual monthly tariff net sums without reductions:
\describe{
   \item{GEMA}{GEMA net tariff.}
   \item{GVL}{GVL net tariff.}
 }}

\item{\code{reduction}}{A named list of applicable reductions for the montly GEMA net tariff:
\describe{
   \item{charity}{The amount of reduction applicable for charity.}
   \item{rcs}{The amount of reduction applicable for religious,
      cultural or social character.}
   \item{promoYoung}{The amount of reduction applicable for promoting young musicians.}
   \item{yearly}{The amount of reduction applicable for a yearly contrat in place.}
   \item{launch}{The amount of reduction applicable (applying \code{launch}).}
   \item{discount}{Other global discount ("Gesamtvertragsnachlass").}
 }}

\item{\code{invoice}}{A named list containing the final invoice sums:
\describe{
   \item{GEMA}{Final GEMA net tariff.}
   \item{GVL}{Final GVL net tariff.}
   \item{net}{The net sum.}
   \item{VAT}{The applicable VAT tax amount.}
   \item{gross}{The gross sum.}
 }}
}}

\keyword{classes}
