% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_class_06_GEMA.invoice_u_t.R
\docType{class}
\name{GEMA.invoice_u_t,-class}
\alias{GEMA.invoice_u_t,-class}
\alias{GEMA.invoice_u_t-class}
\title{S4 Class GEMA.invoice_u_t}
\description{
This class is used for objects that are returned by methods like \code{\link[GEMATariff:GEMA_u_t]{GEMA_u_t}}.
}
\section{Slots}{

\describe{
\item{\code{factors}}{A named list of factors that affect the end result:
\describe{
  \item{GVL}{Fees for GVL.}
  \item{VAT}{Value added tax.}
}}

\item{\code{categories}}{A named list of price categories:
\describe{
  \item{space}{Space category (in 100 square meter steps).}
  \item{admission}{Admission category (in 2€ steps).}
}}

\item{\code{liveclub}}{An object of class \code{\link[GEMATariff]{GEMA.liveclub-class}}.}

\item{\code{tariff}}{A named list containing the actual monthly tariff net sums without reductions:
\describe{
   \item{GEMA}{GEMA net tariff.}
   \item{GVL}{GVL net tariff.}
 }}

\item{\code{reduction}}{A named list of applicable reductions for the GEMA net tariff:
\describe{
   \item{discount}{Other global discount ("Gesamtvertragsnachlass").}
 }}

\item{\code{invoice}}{A named list containing the final invoice sums:
\describe{
   \item{GEMA}{Final GEMA net tariff.}
   \item{GVL}{Final GVL net tariff.}
   \item{net}{The net sum.}
   \item{VAT}{The applicable VAT tax amount.}
   \item{gross}{The gross sum.}
 }}
}}
\keyword{classes}

