% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_class_02_GEMA.liveclub.R
\docType{class}
\name{GEMA.liveclub,-class}
\alias{GEMA.liveclub,-class}
\alias{GEMA.liveclub-class}
\title{S4 Class GEMA.liveclub}
\description{
This class is used for objects that define one live club.
}
\details{
Use function \code{\link[GEMATariff:liveclub]{liveclub}} to create objects of this class.
}
\section{Slots}{

\describe{
\item{\code{space}}{Numeric, room size for the live club in square meters.}

\item{\code{days}}{Integer number between 1 and 30,
      number of days the club is opened per month.}

\item{\code{admission}}{Numeric,
      highest ticket price for the live club. If prices vary over the week,
use the average of highest prices, respectively.}

\item{\code{plusMedium}}{Logical,
      whether in addition to the live music there is also music from
physical recording media (e.g., CDs) being played.}
}}
\examples{
myLiveClub <- liveclub(space=200, days=3, admission=6)
}
\keyword{classes}

