% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEMATariff-package.R
\docType{package}
\name{GEMATariff-package}
\alias{GEMATariff-package}
\title{The GEMATariff Package}
\description{
Calculate GEMA tariffs.
}
\details{
\tabular{ll}{
Package: \tab GEMATariff\cr
Type: \tab Package\cr
Version: \tab 0.01-0\cr
Date: \tab 2016-09-10\cr
Depends: \tab R (>= 2.9.0),methods\cr
Encoding: \tab UTF-8\cr
License: \tab AGPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab https://www.c3s.cc\cr
}

This R package was developed to get a better understanding of the tariff system of the
German collective rights society GEMA. It is intended to be useful as a research tool, and
not for any practical application beyond that. Some of the tariffs are quite complex.
Allthough we implemented everything to the best of our abilities, we do not warranty the
correctness of these calculations. Also, the tariffs are constantly subject to change,
      and while
you read this, some of these changes might not have been implemented yet. If you need
reliable feed back on these tariffs, you should try the official GEMA online calculator
(https://online.gema.de/aidaos/) instead.
}
\author{
m.eik michalke
}
\keyword{package}

