% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gig.R
\name{gig}
\alias{gig}
\title{Define a live concert}
\usage{
gig(gross = 0, guests = 0, admission = 0, musicians = TRUE,
  addIncome = FALSE, fee = 0, production = 0, charity = FALSE,
  rcs = FALSE, promoYoung = FALSE, lumpSumContract = 0,
  plusMedium = FALSE)
}
\arguments{
\item{gross}{Numeric, the total gross receipts for the concert.}

\item{guests}{Numeric, number of people attending the concert.}

\item{admission}{Numeric, highest ticket price for the concert.}

\item{musicians}{Logical,
      whether the music ist mostly perfomed by live musicians (in contrast to
mostly by using physical recording media).}

\item{addIncome}{Logical, whether you have additional income through advertising etc.}

\item{fee}{Numeric, the amount of money payed to the artists,
      relevant if this is a concert
for a strictly limited autience or public without any admission fee.}

\item{production}{Numeric, the directly associated production costs,
      relevant if \code{fee} is relevant, too.}

\item{charity}{Logical, whether the event is of non-profit/charitable character.}

\item{rcs}{Logical, whether the event is of religious, cultural, or social character.}

\item{promoYoung}{Logical, whether this event is only promoting young musicians.}

\item{lumpSumContract}{Numeric,
      one of the following to declare it you have a yearly contract:
\describe{
  \item{0}{No yearly contract or not more than 15 events a year.}
  \item{16}{Yearly contract and at least 16 events a year.}
  \item{31}{Yearly contract and at least 31 events a year.}
}}

\item{plusMedium}{Logical, whether in addition to the live music there is also music from
physical recording media (e.g., CDs) being played.}
}
\value{
An object of class \code{\link[GEMATariff]{GEMA.gig-class}}.
}
\description{
There are two different modes to calculate the tariff for concerts. If you provide the exact
gross revenue rained through the concert,
      that amount is used as the calculation basis. If not,
the product of ticket price and number of guests will be used as an estimate.
}
\examples{
myGig <- gig(guests=100, admission=5)
}
