% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liveclub.R
\name{liveclub}
\alias{liveclub}
\title{Define a live club}
\usage{
liveclub(space = 100, days = 1, admission = 0, plusMedium = FALSE)
}
\arguments{
\item{space}{Numeric, room size for the live club in square meters.}

\item{days}{Integer number between 1 and 30, number of days the club is opened per month.}

\item{admission}{Numeric,
      highest ticket price for the live club. If prices vary over the week,
use the average of highest prices, respectively.}

\item{plusMedium}{Logical, whether in addition to the live music there is also music from
physical recording media (e.g., CDs) being played.}
}
\value{
An object of class \code{\link[GEMATariff]{GEMA.liveclub-class}}.
}
\description{
Define a live club
}
\examples{
myLiveClub <- liveclub(space=200, days=20, admission=6, plusMedium=TRUE)
}
