% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_method_04_plot.R
\docType{methods}
\name{plot,GEMA.dancefloor,missing-method}
\alias{plot,-methods}
\alias{plot,GEMA.dancefloor,missing-method}
\alias{plot,GEMA.dancefloor-method}
\alias{plot,GEMA.gig,missing-method}
\alias{plot,GEMA.gig-method}
\title{Plot method for objects of class GEMA.dancefloor}
\usage{
\S4method{plot}{GEMA.dancefloor,missing}(x,
  year = as.numeric(format(Sys.time(), "\%Y")), xlab = "Year",
  ylab = "Tariff (€)", main = "Development of the GEMA tariff", ...)

\S4method{plot}{GEMA.gig,missing}(x, year = as.numeric(format(Sys.time(),
  "\%Y")), xlab = "Year", ylab = "Tariff (€)",
  main = "Development of the GEMA tariff", ...)
}
\arguments{
\item{x}{An object of class
\code{GEMA.dancefloor} or
\code{GEMA.gig}.}

\item{year}{Integer between 2014 and 2022, draws a line at the given year.}

\item{xlab}{See \code{\link[graphics]{plot}}.}

\item{ylab}{See \code{\link[graphics]{plot}}.}

\item{main}{See \code{\link[graphics]{plot}}.}

\item{...}{Additional options given to \code{plot}.}

\item{y}{From the generic \code{plot} function, ignored.}
}
\description{
Plot method for S4 objects of classes of the \code{GEMATariff} package,
plots the development of the new club tariff from 2014 to 2022.
}
\keyword{methods}
\keyword{plot}

