% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_method_01_show.R
\docType{methods}
\name{show,GEMA.dancefloor-method}
\alias{show,GEMA.dancefloor-method}
\alias{show,-methods}
\alias{show,GEMA.liveclub-method}
\alias{show,GEMA.gig-method}
\alias{show,GEMA.invoice_m_cd-method}
\alias{show,GEMA.invoice_u_t-method}
\alias{show,GEMA.invoice_u_k-method}
\title{Show methods for GEMATariff objects}
\usage{
\S4method{show}{GEMA.dancefloor}(object)

\S4method{show}{GEMA.liveclub}(object)

\S4method{show}{GEMA.gig}(object)

\S4method{show}{GEMA.invoice_m_cd}(object)

\S4method{show}{GEMA.invoice_u_t}(object)

\S4method{show}{GEMA.invoice_u_k}(object)
}
\arguments{
\item{object}{An object of class
\code{GEMA.dancefloor},
\code{GEMA.liveclub},
\code{GEMA.gig},
\code{GEMA.invoice_u_t},
 or
\code{GEMA.invoice_m_cd}.}
}
\description{
Show methods for GEMATariff objects
}
\examples{
\dontrun{
MTLD(tagged.txt)
}
}
