% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_method_02_GEMA_m_cd.R,
%   R/02_method_03_GEMA_m_u_iii.R, R/02_method_05_GEMA_u_t.R,
%   R/02_method_06_GEMA_u_k.R
\docType{methods}
\name{GEMA_m_cd}
\alias{GEMA_m_cd}
\alias{GEMA_m_cd,GEMA.dancefloor-method}
\alias{GEMA_m_cd}
\alias{GEMA_m_u_iii}
\alias{GEMA_m_u_iii,GEMA.dancefloor-method}
\alias{GEMA_m_u_iii}
\alias{GEMA_u_t}
\alias{GEMA_u_t,GEMA.liveclub-method}
\alias{GEMA_u_t}
\alias{GEMA_u_k}
\alias{GEMA_u_k,GEMA.gig-method}
\alias{GEMA_u_k}
\title{S4 methods to calculate tariffs for given dancefloors}
\usage{
GEMA_m_cd(dancefloor, year = as.numeric(format(Sys.time(), "\%Y")),
  discount = FALSE)

\S4method{GEMA_m_cd}{GEMA.dancefloor}(dancefloor,
  year = as.numeric(format(Sys.time(), "\%Y")), discount = FALSE)

GEMA_m_u_iii(dancefloor, discount = FALSE)

\S4method{GEMA_m_u_iii}{GEMA.dancefloor}(dancefloor, discount = FALSE)

GEMA_u_t(liveclub, discount = FALSE)

\S4method{GEMA_u_t}{GEMA.liveclub}(liveclub, discount = FALSE)

GEMA_u_k(gig, byGross = TRUE, year = as.numeric(format(Sys.time(), "\%Y")),
  discount = FALSE)

\S4method{GEMA_u_k}{GEMA.gig}(gig, byGross = TRUE,
  year = as.numeric(format(Sys.time(), "\%Y")), discount = FALSE)
}
\arguments{
\item{year}{Integer. If between 2014 and 2021,
      the market entry reduction will be calculated.
Set to a different year to disable.}

\item{discount}{Logical,
      whether a reduction due to a global contract ("Gesamtvertrag") is eligable. As of now, a reduction of 20\%
is granted if this is set to \code{TRUE}.}

\item{byGross}{Locical,
      whether calculations should be based on \code{gross} value instead of code{guests} \times{} \code{admission}.}

\item{obj}{An object of class \code{\link[GEMATariff]{GEMA.dancefloor-class}}.}

\item{obj}{An object of class \code{\link[GEMATariff]{GEMA.dancefloor-class}}.}

\item{discount}{Logical,
      whether a reduction due to a global contract ("Gesamtvertrag") is eligable. As of now, a reduction of 20\%
is granted if this is set to \code{TRUE}.}

\item{obj}{An object of class \code{\link[GEMATariff]{GEMA.liveclub-class}}.}

\item{discount}{Logical,
      whether a reduction due to a global contract ("Gesamtvertrag") is eligable. As of now, a reduction of 20\%
is granted if this is set to \code{TRUE}.}

\item{obj}{An object of class \code{\link[GEMATariff]{GEMA.gig-class}}.}

\item{year}{Integer. If between 2015 and 2022,
      the market entry reduction will be calculated.
Set to a different year to disable.}

\item{discount}{Logical,
      whether a reduction due to a global contract ("Gesamtvertrag") is eligable. As of now, a reduction of 20\%
is granted if this is set to \code{TRUE}.}
}
\description{
The method \code{GEMA_m_u_iii} calculates GEMA tarifs as of 2012,
      \code{GEMA_m_cd} calculates
the new tariffs as of 2014.

Calculate GEMA tariff M U-III

Calculate GEMA tariff U-T I (Musikaufführungen in Tanzlokalen)

Calculate GEMA tariff U-K I (Wiedergaben mit Musikern)
}
\references{
\url{https://www.gema.de/fileadmin/user_upload/Musiknutzer/Tarife/Tarife_AD/tarif_m_cd.pdf}

\url{https://www.gema.de/fileadmin/user_upload/Musiknutzer/Tarife/Tarife_ad/tarif_u_t.pdf}

\url{https://www.gema.de/fileadmin/user_upload/Musiknutzer/Tarife/Tarife_ad/tarif_u_k.pdf}
}
